기능: 레시피 등록 페이지 개발
* Validation 체크 구현
* 공통 js 분리
* 공통 이미지 경로 추가
* DTO 네이밍 스타일 변경
* 테스트 수정

해결: #12

+++ src/main/java/io/myrecipes/front/controller/IndexController.java
@@ -39,7 +39,7 @@ public class IndexController {
         PageParam pageParamForNewList = new PageParam(1, this.pageSize, this.sortField, this.isDescending);
         List<Recipe> newRecipeList = this.indexService.readRecipeList(pageParamForNewList);
 
-        long recipePageCnt = this.indexService.readRecipePageCnt();
+        int recipePageCnt = this.indexService.readRecipePageCnt();
 
         model.addAttribute("popularRecipeList", popularRecipeList);
         model.addAttribute("newRecipeList", newRecipeList);
